% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm0012_equations.R
\name{calculate_leakage_emissions_acm0012}
\alias{calculate_leakage_emissions_acm0012}
\title{Calculate leakage emissions for ACM0012}
\usage{
calculate_leakage_emissions_acm0012(leakage_energy_mwh, leakage_ef_t_per_mwh)
}
\arguments{
\item{leakage_energy_mwh}{Energy subject to leakage effects in megawatt-hours.}

\item{leakage_ef_t_per_mwh}{Leakage emission factor in tonnes CO2e per
megawatt-hour.}
}
\value{
Numeric vector of leakage emissions in tonnes CO2e.
}
\description{
Accounts for leakage associated with upstream fuel supply displacement and
downstream effects of exported energy.
}
\examples{
calculate_leakage_emissions_acm0012(20, 0.1)
}
